Abstract JsonDelimitedArrayConverter
